// src/services/loopd/protos/loop_pb.d.ts
export class LoopOutRequest extends jspb.Message {
  getAmt(): number;
  setAmt(value: number): void;

  getDest(): string;
  setDest(value: string): void;

  getMaxSwapRoutingFee(): number;
  setMaxSwapRoutingFee(value: number): void;

  getMaxPrepayRoutingFee(): number;
  setMaxPrepayRoutingFee(value: number): void;

  getMaxSwapFee(): number;
  setMaxSwapFee(value: number): void;

  getMaxPrepayAmt(): number;
  setMaxPrepayAmt(value: number): void;

  getMaxMinerFee(): number;
  setMaxMinerFee(value: number): void;

  getLoopOutChannel(): number;
  setLoopOutChannel(value: number): void;

  clearOutgoingChanSetList(): void;
  getOutgoingChanSetList(): Array<number>;
  setOutgoingChanSetList(value: Array<number>): void;
  addOutgoingChanSet(value: number, index?: number): number;

  getSweepConfTarget(): number;
  setSweepConfTarget(value: number): void;

  getHtlcConfirmations(): number;
  setHtlcConfirmations(value: number): void;

  getSwapPublicationDeadline(): number;
  setSwapPublicationDeadline(value: number): void;

  getLabel(): string;
  setLabel(value: string): void;

  getInitiator(): string;
  setInitiator(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): LoopOutRequest.AsObject;
  static toObject(includeInstance: boolean, msg: LoopOutRequest): LoopOutRequest.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: LoopOutRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): LoopOutRequest;
  static deserializeBinaryFromReader(message: LoopOutRequest, reader: jspb.BinaryReader): LoopOutRequest;
}

export namespace LoopOutRequest {
  export type AsObject = {
    amt: number,
    dest: string,
    maxSwapRoutingFee: number,
    maxPrepayRoutingFee: number,
    maxSwapFee: number,
    maxPrepayAmt: number,
    maxMinerFee: number,
    loopOutChannel: number,
    outgoingChanSetList: Array<number>,
    sweepConfTarget: number,
    htlcConfirmations: number,
    swapPublicationDeadline: number,
    label: string,
    initiator: string,
  }
}
// ...
export class SwapResponse extends jspb.Message {
  getId(): string;
  setId(value: string): void;

  getIdBytes(): Uint8Array | string;
  getIdBytes_asU8(): Uint8Array;
  getIdBytes_asB64(): string;
  setIdBytes(value: Uint8Array | string): void;

  getHtlcAddress(): string;
  setHtlcAddress(value: string): void;

  getHtlcAddressNp2wsh(): string;
  setHtlcAddressNp2wsh(value: string): void;

  getHtlcAddressP2wsh(): string;
  setHtlcAddressP2wsh(value: string): void;

  getServerMessage(): string;
  setServerMessage(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): SwapResponse.AsObject;
  static toObject(includeInstance: boolean, msg: SwapResponse): SwapResponse.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: SwapResponse, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): SwapResponse;
  static deserializeBinaryFromReader(message: SwapResponse, reader: jspb.BinaryReader): SwapResponse;
}

export namespace SwapResponse {
  export type AsObject = {
    id: string,
    idBytes: Uint8Array | string,
    htlcAddress: string,
    htlcAddressNp2wsh: string,
    htlcAddressP2wsh: string,
    serverMessage: string,
  }
}

export class MonitorRequest extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MonitorRequest.AsObject;
  static toObject(includeInstance: boolean, msg: MonitorRequest): MonitorRequest.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: MonitorRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MonitorRequest;
  static deserializeBinaryFromReader(message: MonitorRequest, reader: jspb.BinaryReader): MonitorRequest;
}

export namespace MonitorRequest {
  export type AsObject = {
  }
}

export class SwapStatus extends jspb.Message {
  getAmt(): number;
  setAmt(value: number): void;

  getId(): string;
  setId(value: string): void;

  getIdBytes(): Uint8Array | string;
  getIdBytes_asU8(): Uint8Array;
  getIdBytes_asB64(): string;
  setIdBytes(value: Uint8Array | string): void;

  getType(): SwapTypeMap[keyof SwapTypeMap];
  setType(value: SwapTypeMap[keyof SwapTypeMap]): void;

  getState(): SwapStateMap[keyof SwapStateMap];
  setState(value: SwapStateMap[keyof SwapStateMap]): void;

  getFailureReason(): FailureReasonMap[keyof FailureReasonMap];
  setFailureReason(value: FailureReasonMap[keyof FailureReasonMap]): void;

  getInitiationTime(): number;
  setInitiationTime(value: number): void;

  getLastUpdateTime(): number;
  setLastUpdateTime(value: number): void;

  getHtlcAddress(): string;
  setHtlcAddress(value: string): void;

  getHtlcAddressP2wsh(): string;
  setHtlcAddressP2wsh(value: string): void;

  getHtlcAddressNp2wsh(): string;
  setHtlcAddressNp2wsh(value: string): void;

  getCostServer(): number;
  setCostServer(value: number): void;

  getCostOnchain(): number;
  setCostOnchain(value: number): void;

  getCostOffchain(): number;
  setCostOffchain(value: number): void;

  getLastHop(): Uint8Array | string;
  getLastHop_asU8(): Uint8Array;
  getLastHop_asB64(): string;
  setLastHop(value: Uint8Array | string): void;

  clearOutgoingChanSetList(): void;
  getOutgoingChanSetList(): Array<number>;
  setOutgoingChanSetList(value: Array<number>): void;
  addOutgoingChanSet(value: number, index?: number): number;

  getLabel(): string;
  setLabel(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): SwapStatus.AsObject;
  static toObject(includeInstance: boolean, msg: SwapStatus): SwapStatus.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: SwapStatus, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): SwapStatus;
  static deserializeBinaryFromReader(message: SwapStatus, reader: jspb.BinaryReader): SwapStatus;
}

export namespace SwapStatus {
  export type AsObject = {
    amt: number,
    id: string,
    idBytes: Uint8Array | string,
    type: SwapTypeMap[keyof SwapTypeMap],
    state: SwapStateMap[keyof SwapStateMap],
    failureReason: FailureReasonMap[keyof FailureReasonMap],
    initiationTime: number,
    lastUpdateTime: number,
    htlcAddress: string,
    htlcAddressP2wsh: string,
    htlcAddressNp2wsh: string,
    costServer: number,
    costOnchain: number,
    costOffchain: number,
    lastHop: Uint8Array | string,
    outgoingChanSetList: Array<number>,
    label: string,
  }
}
// ...
export namespace TermsRequest {
  export type AsObject = {
  }
}
// ...
export namespace OutTermsResponse {
  export type AsObject = {
    minSwapAmount: number,
    maxSwapAmount: number,
    minCltvDelta: number,
    maxCltvDelta: number,
  }
}

export class QuoteRequest extends jspb.Message {
  getAmt(): number;
  setAmt(value: number): void;

  getConfTarget(): number;
  setConfTarget(value: number): void;

  getExternalHtlc(): boolean;
  setExternalHtlc(value: boolean): void;

  getSwapPublicationDeadline(): number;
  setSwapPublicationDeadline(value: number): void;

  getLoopInLastHop(): Uint8Array | string;
  getLoopInLastHop_asU8(): Uint8Array;
  getLoopInLastHop_asB64(): string;
  setLoopInLastHop(value: Uint8Array | string): void;

  getPrivate(): boolean;
  setPrivate(value: boolean): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): QuoteRequest.AsObject;
  static toObject(includeInstance: boolean, msg: QuoteRequest): QuoteRequest.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: QuoteRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): QuoteRequest;
  static deserializeBinaryFromReader(message: QuoteRequest, reader: jspb.BinaryReader): QuoteRequest;
}

export namespace QuoteRequest {
  export type AsObject = {
    amt: number,
    confTarget: number,
    externalHtlc: boolean,
    swapPublicationDeadline: number,
    loopInLastHop: Uint8Array | string,
    pb_private: boolean,
  }
}
// ...
export class OutQuoteResponse extends jspb.Message {
  getSwapFeeSat(): number;
  setSwapFeeSat(value: number): void;

  getPrepayAmtSat(): number;
  setPrepayAmtSat(value: number): void;

  getHtlcSweepFeeSat(): number;
  setHtlcSweepFeeSat(value: number): void;

  getSwapPaymentDest(): Uint8Array | string;
  getSwapPaymentDest_asU8(): Uint8Array;
  getSwapPaymentDest_asB64(): string;
  setSwapPaymentDest(value: Uint8Array | string): void;

  getCltvDelta(): number;
  setCltvDelta(value: number): void;

  getConfTarget(): number;
  setConfTarget(value: number): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): OutQuoteResponse.AsObject;
  static toObject(includeInstance: boolean, msg: OutQuoteResponse): OutQuoteResponse.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: OutQuoteResponse, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): OutQuoteResponse;
  static deserializeBinaryFromReader(message: OutQuoteResponse, reader: jspb.BinaryReader): OutQuoteResponse;
}

export namespace OutQuoteResponse {
  export type AsObject = {
    swapFeeSat: number,
    prepayAmtSat: number,
    htlcSweepFeeSat: number,
    swapPaymentDest: Uint8Array | string,
    cltvDelta: number,
    confTarget: number,
  }
}
// ...
export class TokensRequest extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TokensRequest.AsObject;
  static toObject(includeInstance: boolean, msg: TokensRequest): TokensRequest.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TokensRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TokensRequest;
  static deserializeBinaryFromReader(message: TokensRequest, reader: jspb.BinaryReader): TokensRequest;
}

export namespace TokensRequest {
  export type AsObject = {
  }
}

export class TokensResponse extends jspb.Message {
  clearTokensList(): void;
  getTokensList(): Array<LsatToken>;
  setTokensList(value: Array<LsatToken>): void;
  addTokens(value?: LsatToken, index?: number): LsatToken;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TokensResponse.AsObject;
  static toObject(includeInstance: boolean, msg: TokensResponse): TokensResponse.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TokensResponse, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TokensResponse;
  static deserializeBinaryFromReader(message: TokensResponse, reader: jspb.BinaryReader): TokensResponse;
}

export namespace TokensResponse {
  export type AsObject = {
    tokensList: Array<LsatToken.AsObject>,
  }
}
// ...
export const SwapState: SwapStateMap;

export interface FailureReasonMap {
  FAILURE_REASON_NONE: 0;
  FAILURE_REASON_OFFCHAIN: 1;
  FAILURE_REASON_TIMEOUT: 2;
  FAILURE_REASON_SWEEP_TIMEOUT: 3;
  FAILURE_REASON_INSUFFICIENT_VALUE: 4;
  FAILURE_REASON_TEMPORARY: 5;
  FAILURE_REASON_INCORRECT_AMOUNT: 6;
}

export const FailureReason: FailureReasonMap;